Enhanced the OLE2 mutator multiple mutator
If the directory with samples selected for the projects contains too many other type of files, it often sends back an empty ZIP file with no mutated OLE2 document at all. There is very little we can do here but, at the very least, we can try to iterate 2 times the number of files the user wants to generate before giving up.