Added the option to ignore duplicate crashes
NFP: Generate a crash hash for each crash based on the last 3 nibbles of each element in the stack trace up to a maximum of 15 elements. The 1st element in the crash hash are the 3 nibbles of the $PC.
GUI: Added the option to ignore duplicated crashes at project level. Configure it via the web GUI in Projects. Disabled by default.