Many little changes
MINIMIZER: All the code of the minimizer have been split into pieces, in order to make it more maintainable.
GDB: Add back the "shell = True" for the GDB debugging interface.
NFP: Do not log duplicate crashes if the option to discard duplicates is set even for exploitable ones.
FRONTEND: Added the option to search by crash_hash in the GUI.